test: 예약 서비스 예외 테스트 추가
